xyssl-0.3

+++ src/md2.c
@@ -155,7 +155,7 @@ int md2_file( char *filename, uchar digest[16] )
     md2_starts( &ctx );
 
     while( ( n = fread( buf, 1, sizeof( buf ), f ) ) > 0 )
-        md2_update( &ctx, buf, n );
+        md2_update( &ctx, buf, (uint) n );
 
     md2_finish( &ctx, digest );
