+++ src/md5.c
@@ -279,7 +279,7 @@ int md5_file( char *filename, uchar digest[16] )
     md5_starts( &ctx );
 
     while( ( n = fread( buf, 1, sizeof( buf ), f ) ) > 0 )
-        md5_update( &ctx, buf, n );
+        md5_update( &ctx, buf, (uint) n );
 
     md5_finish( &ctx, digest );
