+++ src/sha1.c
@@ -314,7 +314,7 @@ int sha1_file( char *filename, uchar digest[20] )
     sha1_starts( &ctx );
 
     while( ( n = fread( buf, 1, sizeof( buf ), f ) ) > 0 )
-        sha1_update( &ctx, buf, n );
+        sha1_update( &ctx, buf, (uint) n );
 
     sha1_finish( &ctx, digest );
